+++ .storybook/decorators/withGlobalEvent.tsx
@@ -1,5 +1,0 @@
-import { GlobalEventProvider } from '@vibrant-ui/core';
-
-export function withGlobalEvent(storyFn) {
-  return <GlobalEventProvider>{storyFn()}</GlobalEventProvider>;
-}