+++ packages/vibrant-components/src/lib/Dropdown/Dropdown.tsx
@@ -1,18 +1,19 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import {
   Box,
+  OverlayBox,
   ScrollBox,
   ThemeProvider,
   getWindowDimensions,
   useCurrentThemeMode,
   useResponsiveValue,
+  useSafeArea,
   useWindowDimensions,
 } from '@vibrant-ui/core';
 import { Transition } from '@vibrant-ui/motion';
 import { detectOverflow, flipPosition, getElementRect, getOffsetByPosition } from '@vibrant-ui/utils';
 import type { LayoutEvent, Position, Rect } from '@vibrant-ui/utils';
 import { Backdrop } from '../Backdrop';
-import { Dismissible } from '../Dismissible';
 import { withDropdownVariation } from './DropdownProps';
 
 const CONTENT_PADDING = 20;
@@ -75,6 +76,7 @@ export const Dropdown = withDropdownVariation(
     const [offset, setOffset] = useState<{ x?: number; y?: number }>({});
     const [contentHeight, setContentHeight] = useState<number>();
     const { height: viewportHeight } = useWindowDimensions();
+    const { insets } = useSafeArea();
 
     const { breakpointIndex } = useResponsiveValue({ rootBreakPoints: true });
     const isMobile = breakpointIndex === 0;
@@ -110,15 +112,15 @@ export const Dropdown = withDropdownVariation(
     }, []);
 
     const handleContentResize = useCallback(
-      async ({ layout: { width, height, x, y } }: LayoutEvent) => {
+      async ({ width, height, top, left }: LayoutEvent) => {
         if (!isMobile) {
           const openerRect = await getElementRect(openerRef.current);
 
           const { x: offsetX, y: offsetY } = getOffsetAvoidingOverflowByPosition(
             openerRect,
             {
-              x,
-              y,
+              x: left,
+              y: top,
               width,
               height: height + CONTENT_PADDING * 2,
             },
@@ -151,54 +153,48 @@ export const Dropdown = withDropdownVariation(
         <Box ref={openerRef}>{opener}</Box>
         {isOpen && !isMobile && (
           <ThemeProvider theme={rootThemeMode}>
-            <Dismissible active={visible} onDismiss={closeDropdown}>
-              <Transition
-                ref={targetRef}
-                animation={{
-                  opacity: visible ? 1 : 0,
-                  x: offset.x,
-                  y: offset.y,
-                }}
-                style={{
-                  x: offset.x,
-                  y: offset.y,
-                }}
-                duration={150}
-              >
-                <Box position="absolute" zIndex={Z_INDEX}>
-                  <Box
-                    backgroundColor="background"
-                    py={CONTENT_PADDING}
-                    elevationLevel={4}
-                    borderRadiusLevel={1}
-                    width={[280, 280, 240]}
+            <Transition
+              animation={{
+                opacity: visible ? 1 : 0,
+                x: offset.x,
+                y: offset.y,
+              }}
+              style={{
+                x: offset.x,
+                y: offset.y,
+              }}
+              duration={150}
+            >
+              <OverlayBox open={true} ref={targetRef} onDismiss={closeDropdown} targetRef={openerRef} zIndex={Z_INDEX}>
+                <Box
+                  backgroundColor="surface2"
+                  py={CONTENT_PADDING}
+                  elevationLevel={4}
+                  borderRadiusLevel={1}
+                  width={[280, 280, 240]}
+                >
+                  <Transition
+                    animation={
+                      visible
+                        ? {
+                            height: contentHeight,
+                          }
+                        : {}
+                    }
+                    duration={150}
                   >
-                    <Transition
-                      animation={
-                        visible
-                          ? {
-                              height: contentHeight,
-                            }
-                          : {}
-                      }
-                      style={{
-                        height: contentHeight,
-                      }}
-                      duration={150}
-                    >
-                      <Box overflow="hidden">
-                        <Box onLayout={handleContentResize} flexShrink={0}>
-                          {renderContents(closeDropdown)}
-                        </Box>
+                    <Box overflow="hidden" height={contentHeight}>
+                      <Box onLayout={handleContentResize} flexShrink={0}>
+                        {renderContents(closeDropdown)}
                       </Box>
-                    </Transition>
-                  </Box>
+                    </Box>
+                  </Transition>
                 </Box>
-              </Transition>
-            </Dismissible>
+              </OverlayBox>
+            </Transition>
           </ThemeProvider>
         )}
-        {isOpen && isMobile && (
+        {isMobile && (
           <ThemeProvider theme={rootThemeMode}>
             <Backdrop open={isOpen} zIndex={Z_INDEX} onClick={closeDropdown} transitionDuration={visible ? 150 : 100}>
               <Transition
@@ -215,7 +211,7 @@ export const Dropdown = withDropdownVariation(
                   pb={BOTTOM_SHEET_CONTENT_BOTTOM_PADDING}
                   width="100%"
                   maxHeight={viewportHeight - 120}
-                  backgroundColor="background"
+                  backgroundColor="surface2"
                   borderTopLeftRadiusLevel={4}
                   borderTopRightRadiusLevel={4}
                 >
@@ -226,6 +222,7 @@ export const Dropdown = withDropdownVariation(
                     duration={150}
                   >
                     <ScrollBox
+                      height={contentHeight}
                       hideScroll={
                         (contentHeight ?? 0) +
                           (BOTTOM_SHEET_CONTENT_TOP_PADDING + BOTTOM_SHEET_CONTENT_BOTTOM_PADDING) <=
@@ -233,7 +230,7 @@ export const Dropdown = withDropdownVariation(
                       }
                     >
                       <Box onLayout={handleContentResize} flexShrink={0}>
-                        {renderContents(closeDropdown)}
+                        <Box pb={insets.bottom}>{renderContents(closeDropdown)}</Box>
                       </Box>
                     </ScrollBox>
                   </Transition>