+++ packages/vibrant-utils/src/lib/type/DistributivePick/DistributivePick.ts
@@ -1,0 +1,1 @@
+export type DistributivePick<Type, Key extends keyof Type> = Type extends unknown ? Pick<Type, Key> : never;